Renamed routes.jsx and main.jsx so things make more sense
